fix code comments and rename file
